restructure project for unit tests
